Productionization_flight_price_prediction/requirements.txt
asttokens==2.4.1
backcall==0.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
Flask==3.0.3
fonttools==4.53.1
imageio==2.34.2
ipykernel==5.5.6
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.1.0
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sklearn-pandas==2.2.0
stack-data==0.6.3
threadpoolctl==3.5.0
tifffile==2024.7.24
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.3
xgboost==2.1.1
